AbstractTypeAdapter and stub methods to TypeAdapterInterface
