+++ src/Ifsnop/Mysqldump/TypeAdapter/TypeAdapterPgsql.php
@@ -2,6 +2,6 @@
 
 namespace Ifsnop\Mysqldump\TypeAdapter;
 
-class TypeAdapterPgsql extends TypeAdapterFactory implements TypeAdapterInterface
+class TypeAdapterPgsql extends AbstractTypeAdapter implements TypeAdapterInterface
 {
 }
